Build the raytracing simulation using python
This implements the simulation of seismic rays and gives some ways to
display the results.
Whenever a ray hits the surface it's position and the time it took the
ray are recorded in the Detector from where the data can be processed
and displayed at a later time.
The Model holds the definition of the different layers.
The RayTracer does the heavy lifting by simulating the reflected and
refracted rays.

In the main.py a sample setup is built and displayed nicely.